requirements.txt
certifi==2021.5.30
chardet==4.0.0
gitdb==4.0.7
GitPython==3.1.18
idna==2.10
requests==2.25.1
smmap==4.0.0
urllib3==1.26.5
